requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sqlalchemy>=1.4.52,<2.0
pymysql==1.1.0
pyarrow==12.0.1
python-dotenv==1.0.0
mlflow==2.16.0
pytest==7.4.4
watchdog==3.0.0
PyYAML==6.0.1
requests==2.31.0

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Monitoring and observability
prometheus-client==0.19.0
structlog==24.1.0

# Circuit breaker and resilience
circuitbreaker==1.4.0

# Additional dependencies for production
python-json-logger==2.0.7

# Phase 2: Data quality and feature engineering
scipy==1.10.1

# Phase 3: Vector recall & ranking (optional components)
# Install when USE_FAISS_RECALL=1 or ONNX export is required
# faiss-cpu==1.7.4
# sentence-transformers==2.2.2
lightgbm==4.1.0
# skl2onnx==1.16.0
